requirements.txt
pandas
python-dotenv
tzdata
boto3
selenium
requests
databricks-sql-connector